requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
bcrypt==3.2.0
botocore==1.40.56
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
exceptiongroup==1.3.0
feedparser==6.0.11
future==0.18.2
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-genai==1.46.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperframe==6.1.0
idna==2.10
importlib-metadata==4.6.4
jmespath==1.0.1
oauthlib==3.2.0
packaging==25.0
paramiko==2.9.3
pillow==10.4.0
postgrest==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.11.2
PyJWT==2.9.0
PyNaCl==1.6.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2022.1
realtime==2.5.3
reportlab==4.4.4
requests==2.25.1
rsa==4.9.1
s3transfer==0.13.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.5
websockets==15.0.1
zipp==3.23.0
